Added more consistent validation and some other bug fixes
- favorite toggle now also works on the arcitle edit view
- All places (login(), updateSettings(), signUp(), createArticle(), updateArticle()) shall now have an immediate fluentValidation handling for the incoming model object
  (i.e., form field validation)

Note that if the incoming model in controller methods have an AbstractValidator --> correct 422 and "errors" response (cf. RegisterValidatorsFromAssemblyContaining: automatic ASP.NET validation)
However, if only the command has such an AbstractValidator (ValidationPipelineBehavior), there might be a 422 internally, but that does not produce the correct "errors" response and HTTP status (cf. ErrorHandlingMiddleware).